test: drop tests that cannot fail and split the bundled helper module
